app/requirements/base.txt
Django==2.2
django-appconf==1.0.3
django-compressor==2.2
django-environ==0.4.5
django-recaptcha==2.0.4
django-sass-processor==0.7.3
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
gunicorn==19.9.0
libsass==0.18.0
pytz==2019.1
rcssmin==1.0.6
rjsmin==1.0.12
six==1.12.0
sqlparse==0.3.0
